[54] clear database on test start and refactor
