+++ .github/workflows/rust.yml
@@ -17,6 +17,6 @@ jobs:
     steps:
     - uses: actions/checkout@v4
     - name: Build
-      run: cargo build --verbose
+      run: cargo build --lib --verbose
     - name: Run tests
-      run: cargo test --verbose
+      run: cargo test --lib --verbose
